Toggle LVGL X11 backend via cmake
