+++ lib/lvgl/lv_conf.h
@@ -4,12 +4,24 @@
 #define LV_CONF_INCLUDE_SIMPLE
 
 /* Use SDL by default for development */
+#ifndef LV_USE_SDL
 #define LV_USE_SDL 1
+#endif
+#ifndef LV_USE_X11
 #define LV_USE_X11 0
+#endif
+#ifndef LV_USE_WAYLAND
 #define LV_USE_WAYLAND 0
+#endif
+#ifndef LV_USE_LINUX_FBDEV
 #define LV_USE_LINUX_FBDEV 0
+#endif
+#ifndef LV_USE_NUTTX
 #define LV_USE_NUTTX 0
+#endif
+#ifndef LV_USE_LINUX_DRM
 #define LV_USE_LINUX_DRM 0
+#endif
 #define LV_LINUX_DRM_GBM_BUFFERS 1
 
 #include "src/lv_conf_internal.h"
